requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.6.0
argcomplete==3.2.3
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
curlify==2.2.1
datamodel-code-generator==0.25.5
db-client @ git+https://github.com/katyakorsa/db_client@52e48226201e20cdfefbd78eb0c7b19247f4191b
distconfig3==1.0.1
dm-api-account @ git+https://github.com/katyakorsa/dm_api_account.git@fbbde2946847ea2b71776e02b13fe17c12ac5941
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Faker==24.4.0
genson==1.2.2
greenlet==3.0.3
idna==3.6
inflect==5.6.2
iniconfig==2.0.0
isort==5.13.2
jdcal==1.4.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mimesis==15.1.0
mypy-extensions==1.0.0
openpyxl==2.4.11
orm-client @ git+https://github.com/katyakorsa/orm_client.git@f9658260c2319af637610ca21153da33ac046648
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
PyHamcrest==2.1.0
pytest==7.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
records==0.5.3
requests==2.31.0
rest-client @ git+https://github.com/katyakorsa/rest_client.git@8385b6aedf85bf1788805d8d6103e17c9d35acd7
six==1.16.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.46
structlog==24.1.0
tablib==3.6.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
vyper-config==1.2.1
watchdog==4.0.0
